refactor: modularize Jotai store
